2018-12-23 debug

+++ utilities/nvm_write_cache/prefix_extractor.cc
@@ -58,7 +58,7 @@ ArbitrarilyExtractor::ArbitrarilyExtractor(size_t  range_num)
 
 }
 
-ArbitrarilyExtractor *ArbitrarilyExtractor::NewArbitrarilyExtractor(uint16_t range_num) {
+ArbitrarilyExtractor *ArbitrarilyExtractor::NewArbitrarilyExtractor(size_t range_num) {
     return new ArbitrarilyExtractor(range_num);
 }
